Changed front-end of mp2enc from 16 bits to 32 bits
24 bits are also supported and in fact I only tested this with 24 bit stereo WAV files on Windows (little-endian machine).
Also changed the definition of FLOAT back to "double" as it once was (according to the logs written by my predecessors). Because of this, the code now compiles without warnings.

+++ aenc/common.c
@@ -134,7 +134,7 @@ frame_params *fr_ps;
         bsp = fr_ps->header->bitrate_index;
         br_per_ch = bitrate[lay][bsp] / fr_ps->stereo;
         ws = fr_ps->header->sampling_frequency;
-        sfrq = s_freq[ws];
+        sfrq = (int)s_freq[ws];
         /* decision rules refer to per-channel bitrates (kbits/sec/chan) */
         if ((sfrq == 48 && br_per_ch >= 56) ||
             (br_per_ch >= 56 && br_per_ch <= 80)) table = 0;
